+++ angu8leika/src/app/components/main/main.component.html
@@ -1,8 +1,4 @@
- <!-- ng if either display login or logged -->
+ <!-- ng if either display logged in or logged out -->
     <app-login *ngIf="loggedInState"></app-login>
     <app-logged *ngIf="loggedOutState"></app-logged>
 
-    <!-- <app-login *ngIf="loggedState"></app-login>
-    <app-logged *ngIf="loggedState"></app-logged> -->
-
-<!-- TODO: TEST PROTECTED ROUTE -->
