+++ backend/index.js
@@ -4,6 +4,8 @@ const mongoose = require('mongoose');
 const cors = require('cors');
 // Import Routes
 const authRoute = require('./routes/auth');
+const usersRoute = require('./routes/users');
+
 // cookie parser
 const cookieParser = require('cookie-parser');
 
@@ -46,6 +48,7 @@ app.use(express.urlencoded({
 app.use(cookieParser());
 
 // Routes
-app.use('/api/user', authRoute);
+app.use('/api/auth', authRoute);
+app.use('/api/users', usersRoute);
 
-app.listen(process.env.API_PORT, () => console.log(`Server works: http://localhost:${process.env.API_PORT}/api/user`));
+app.listen(process.env.API_PORT, () => console.log(`Server works: http://localhost:${process.env.API_PORT}/api`));
